Simplify some types based on what is needed
Summary:
I'm continuing to work toward being able to model attribute actions in
a descriptor-like way where an attribute does not have a type but rather
has get and set actions (which are either types, or some sort of indication
of why the action is not permittd).

The next stage of my refactor requires creating an `Attribute` type for
use in `attr.rs`, but given the existing code, in which we've used the
name `Attribute` to indicate the smashing together of a type with the
defining class, this gets very confusing.

The solution is to first note that we are using `defining_class` only
in a few places now, and only internally to `classes.rs`. As a result,
we can:
- make it private to `classes.rs` instead of exposed elsewhere
- rename it to `WithDefiningClass`, so that the name `Attribute` becomes
  available again
- (after observing that `defining_class` is only ever used on instance
  attributes) simplify both the type signature and the implementation
  of `get_class_attribute` a bit by ignoring the defining class, which
  no one cares about anyway

Reviewed By: ndmitchell

Differential Revision: D68687291

fbshipit-source-id: 2cd3751ddb8cf8ceed09ad741ff15b85f8a905cd