Add failing test to show bug with decorator application
Summary:
Currently, we use `bind_key` to store a chain of decorator applications. This interacts poorly with the "anywhere" type.

In the given example, the definition count for `decorated` is 1, so we don't actually update the anywhere binding at all.

For other examples, where the definition count is >1, we would end up with the union of undecorated and decorated signatures (including all intermediate types, in the case of multiple decorators).

The fix in the next diff is to stop modeling decorator application as assignment. This is the simplest fix, leaving "anywhere" binding logic alone.

It might be nice to also make "anywhere" bindings more precise. For example, the second write to `x` in the below example is not reachable form `f`.

```
def f():
  x # read
x = e # write 1 - dead
x = e # write 2
```

Reviewed By: stroxler, ndmitchell

Differential Revision: D68694062

fbshipit-source-id: 6354997df66c659463b3ebe48b7205b14d115495