Cap the number of workers with config option
Summary:
Based on the discussion in [this comment thread](https://fb.workplace.com/groups/pyreqa/posts/9251329598290123/?comment_id=9251652018257881&reply_comment_id=9261631040593312), it seems like there's a good argument for capping the number of workers to avoid Pyre taking too much resources on the host machines.

However, it's extremely unclear to me what would be the best way of calculating the cap. I think it all comes down to this:
- Sometimes we want Pyre to take as much resources it can to minimize turnaround time. This usually happens on CI, and sometimes on devserver / laptop when only Python files are being edited.
- Sometimes Pyre needs to share machines with other resources-intensive tools (Like HHVM, Hack, Buck, etc.), in which case it should really scale down what it wants to consume to avoid hogging CPU/memory.

Problem is that both use cases currently needs to share 1 configuration. And if I tweak the default worker count too much towards one use case, I ended up getting a not-ideal number for the other use case.

The diff currently proposes the simplest, most naive way we could handle this issue: let's simply put a constant cap on the number of workers, when the number is being auto-determined. As long as this number is reasonably large (say, somewhere around 32-64), it should help with edge cases while not slowing down CI checks too much. I've also allowed the cap to be specifiable both with config file and with CLI flag. So in the worst case if we really want to use different settings for CI and for devservers, we could simply add the default to the config file and overriding the cap on CI with CLI flag.

I've also thought about another alternative where we could do some math to determine the cap. E.g. we could estimate total Pyre memory usage by adding up the sharedmem size plus worker * per-worker memory cap (my rough experiments suggests that the per-worker memory usage on the OCaml side is only a few hundreds of megabytes for IG, but [user reports](https://fb.workplace.com/groups/pyreqa/posts/9251329598290123/?comment_id=9251652018257881&reply_comment_id=9263660173723732) a 1.2-1.5GB usage). But then I realize I don't really know what to compare that mem cost against -- should it be 100% of available mem? 50%? 30% 20%? The percentage seems to need to change according to use case as well. And if we are willing to customize different percentages for different use cases, why not letting people customize that cap directly?

Reviewed By: stroxler, tianhan0

Differential Revision: D68584739

fbshipit-source-id: 51e4e640631bdea9dd6c69535e5652eab1546af7