Improve naming for yield bindings
Summary:
In our bindings, we have a binding called `YieldTypeOfYieldAnnotation` which stands for the yield type from an annotation. This is not to be confused with a binding called `YieldTypeOfYield` which infers the yield type based on the yield expressions in the funciton body.

We rename bindings that extract send and return type from annotations to accurately reflect this.

Reviewed By: ndmitchell

Differential Revision: D68727219

fbshipit-source-id: a635284f774930ef29e61777764747929a184b13