Improve check for static and class methods
Summary:
The way we currently check if a callee is a static or class method is pretty hacky, since it just turns the whole expression into a string and tries to use it as a function name.
This is currently leading to performance issues when the expression is large (say, 10k lines, or 300kb).

Let's improve this by delaying the check until we have a proper target, then use `Target.get_module_and_definition` to get the definition and check for decorators.

Reviewed By: tianhan0

Differential Revision: D68633423

fbshipit-source-id: 283e2183b9c3a3a8c9bcd5a54215e19a846ab0d0