Eliminate buggy attempt to validate unpacked targets
Summary:
The existing plumbing where `make_binding` gets called inside `bind_unpacking`
looks elegant but is misleading - it is too elegant: it correctly handles
unidirectional type inference, but as our earlier test showed it is *never*
correct about passing `ann` to the top-level calling code: we wind up trying
to validate that the entire value being assigned to an unpacked target is
valid for every single sub-expression (which is impossible, unless they
are all `tuple[Any]` I suppose).

An ideal solution might work out when it's possible to decompose the target
expression and use bidirectional inference on each sub-expression of the RHS
based on annotations of names in the LHS, but none of that plumbing exists
and it's not obvious that multi-target assignment is common enough to prioritize
this anytime soon.

In the meantime, a pretty reasonable solution is to stick with unidirectional
type inference for the RHS of unpacked assignments by passing `None` to
`make_binding`. This solves our false positives, although we still have a
false negative in that the assignments are *never* checked for whether they
are valid.

I've updated the test cases - which now run with no errors - to indicate
that we have false negatives.

There was an existing test for contextual typing that also encountered this
issue; the text is now error-free but I've left it as `testcase_with_bug`
because I would guess that our v0 solution to the false negative problem won't
use contextual typing.

The false negatives are not good, but are certainly less of a problem than
the always-incorrect checks we had before.

Reviewed By: ndmitchell

Differential Revision: D68764543

fbshipit-source-id: ac95b3ce9525d61587a7e3dc2b2d924fc7c80e97