Wire up CheckAssignExprToAttribute for AnnAssign
Summary:
This diff continues my migration toward using an API that can support
things like setter protocols where an attribute set might, in the most
general case, require checking a function call and invoking things like
the constraint solver / overload resolution.

In this diff, I handle the case of `AnnAssign`, which is a bit easier
than `Assign` because it doesn't require changes to the `bind_target` plumbing.

I'll tackle `Assign` in a separate diff, because that involves dealing
with `bind_target` and preexisting bugs in unpacked target handling, which
is a bit more subtle.

Reviewed By: ndmitchell

Differential Revision: D68767617

fbshipit-source-id: 3f4b26b9b6b3f22a1dd5d9bf4037ed2fa117a6c0