Store the name of model generators in the model
Summary:
# Context
We are having memory issue during model generations, and we suspect that is due to storing all generated models in the ocaml heap and sending them from workers to the main process. To fix this, we want to write those directly to shared memory.

# Problem
We can't easily write generated models to shared memory because of the `--dump-model-query-results`. This option currently needs to have a mapping from model query to all its generated models.

# Solution
As a workaround, let's store the name of the model generator directly in the generated model. This should allow us to write models to shared memory, and still implement `--dump-model-query-results` by iterating on the shared memory.

Reviewed By: tianhan0

Differential Revision: D68721275

fbshipit-source-id: 9a113a01bc332b4bd18c8ddc5cc1bdd9898e52cf