Split the logic of checking base class metaclasses from the logic of computing derived class metaclasses
Summary:
scooby_lets_split_up

There's a function called `calculate_metaclass()`, which one would expect that it just calculates the metaclass for a given class. But that's not the case! The function *both* calculates the metaclass, *and* performs an additional check on that metaclass to make sure it's a subclass of all base class metaclasses. I think we still want the check, but it might also be a good idea to take that check out of the `calculate_metaclass()` function and invoke it separately to avoid the confusion.

One of the reason why the calculation and the check are muddled together was that they need access to one single piece of derived state `base_metaclasses`. But that shared state can easily be computed on the caller side and pass in as input to calculation and check

Reviewed By: rchen152

Differential Revision: D68790969

fbshipit-source-id: 2bf6d60bd0302bead5df0615d635d7ec37780ca9