Use the Result directly for get/set
Summary:
In order to turn `AttributeInner` into an enum that encodes behaviors usefully
once get and set can diverge, we need the downstream code not to be directly
matching on it.

To that end, just hand the inner access `Result<Type, NoAccessReason>` directly
to the lookup code.

There's no behavior change, nor public interface change here - it's entirely some
type tetris in the internal interfaces of `attr.rs`

Reviewed By: ndmitchell

Differential Revision: D68827158

fbshipit-source-id: 96c7f5e8be674e9982c0097a4b47fa800f688314