Turn `AttributeInner` into a variant
Summary:
While I'm at it, get rid of the `impl` for `AttributeInner`, it makes
sense to just flatten this into `Attribute`; the tiny bit of simplicity
from nesting impls isn't worth the extra block of code and the extra
indirection when trying to skim it.

There's no behavior change, the only case right now is the `Simple` case
that wraps a flat access result which we assume applies to both get and set,
and can be computed in advance with a single closed-form type (this won't be
the case for some set actions once we support descriptors and property setters).

But most of the rust-level type plumbing is now in place to start allowing
descriptors where get and set are not equivalent.

Reviewed By: ndmitchell

Differential Revision: D68827159

fbshipit-source-id: 62628ee58a6d3f641d16cd206ba82057fa74f72f