Move ClassField calculation to classes.rs
Summary:
The classes, attr, Attribute, and ClassField code are all tightly coupled in
the sense that implementing anything about descriptor-based behvior has to flow
through a pipeline that touches all of these. Moreover, the details of how ClassField
is represented are *very* private to the classes code.

Having the calculation to produce `ClassField` directly in `answers.rs`'s `solve`
breaks this modularity; it's harder to make changes because they live far apart, and it
forced us to make a few methods public that can really be private.

This commit moves the low-level logic into `classes.rs` where it can live in the same
block of code that deals with class member and attribute representation.

Reviewed By: ndmitchell

Differential Revision: D68828128

fbshipit-source-id: 84eb3710aae8f2e9f158862edd8c75426f372a25