Add a test for classmethod
Summary:
Add an empty class definition to the test stubs (we are going to bake the behavior
in, not treat it as a user-defined descriptor, so the "true" stub isn't relevant)
and add a test that applies it.

Then test shows current behavior: we complain on the decoration because it is
trying to invoke the constructor, and we treat the result as a `classmethod`
instance.

In the next diff I'll add enough plumbing to treat it as a special type, which will
enable us to later build out the correct descriptor behavior inside of attribute
lookups.

(Note: the test is using `reveal_type` rather than `assert_type` because we're going
to have to work with non-user-denotable types as we progress through the implementation;
once it is all working end-to-end we'll be able to validate descriptor behavior directly
and get rid of the `reveal_type`s)

Reviewed By: ndmitchell

Differential Revision: D68825778

fbshipit-source-id: 1f19b00eb652935322604037bee46f07154717f9