Use a first class shared memory table for models in the fixpoint
Summary:
# Context
See D68721275 for context

# This diff
This diff changes the data structure we use for models to `FirstClassWithKeys` (i.e a shared memory table with key tracking), which was introduced in the previous diff.

This required some changes to the API for `FirstClassWithKeys` in order to provide safe map-reduce operations. We allow 3 "modes": ReadOnly, AddOnly (each worker adds new key-value pairs) or PreserveKeyOnly (update existing keys, but do not add any new).

For now, we still produce models from DSL into the ocaml heap using `Registry.t`, and store it all in shared memory right before the taint fixpoint. We will change that in the next diff.

Reviewed By: tianhan0

Differential Revision: D68833107

fbshipit-source-id: d8cd221b4ec864b36690e27738e4e10b44a23482