Move the first class shared memory table with keys
Summary:
# Context
See D68721275 for context

# Why this diff?
We will need a better API to the `SharedModels` module, so we can allow model generation to write models directly in the shared memory using it.

We cannot use `SharedMemory.FirstClass` directly, because we need to remember the set of keys (i.e callables that have an initial model).

Fortunately, we already have a data structure that stores key-value pairs in the shared memory and tracks the set of keys used. This is currently used for caching, so let's make it more generic and move it in the shared memory module.

Reviewed By: tianhan0

Differential Revision: D68832863

fbshipit-source-id: ba1e1579298db8faf710b045757b4c11f0f075ae