Set up a test for staticmethod
Summary:
Add an empty class def to the test stubs (we don't need the real def, we are
going to bake behavior into decorator resolution) and a test.

The test currently shows that we error on the decorator (because it's trying to
invoke the constructor) and that the result of applying the decorator is just
our class definition.

Note: we use `reveal_type` for now rather than `assert_type` because as the
work progresses, we're often going to want to see progress in terms of non-user-denotable
types. Once the descriptor works end-to-end we'll be able to use assert_type to verify
descriptor behavior.

Reviewed By: ndmitchell

Differential Revision: D68892138

fbshipit-source-id: 8e40d2780230d943a69e935cce29d88ad0aae60e