Support staticmethod attribute access
Summary:
This commit adds support for accessing staticmethods as attributes, with the
corect behavior of always skipping binding logic (whether we access off of the
instance or the class).

There is still a gap in functionality for untyped code, because the bindings.rs
code is still assuming that in any method definition the default type of the
first parameter is Self; fixing this problem is orthogonal to my current focus
on attribute access and descriptor behaviors so Im unsure whether Ill be able
to get around to fixing it in this stack.

I left a testcase_with_bug illustrating the gap

Reviewed By: ndmitchell

Differential Revision: D68892137

fbshipit-source-id: ed54971c4b26c4cc13287c28147300e04e098d06