Kill Buck after building targets for infer and analyze command
Summary:
This could release a lot of memory used by Buck.

This diff introduces a flag `kill_buck_after_build` to decide when to kill
Buck (under the `Eager` interface):
- In OCaml, flag `kill_buck_after_build` is stored under `Configuration.Buck`,
  since it is only relevant when using Buck.
- In Pyre client, flag `kill_buck_after_build` is stored under
  `BuckSourcePath`, for the same reason as above.
- In Pyre client, I decided that `kill_buck_after_build` are options for
  subcommand `infer` and `analyze`, since this flag is only relevant for these
  two cases. We may want to kill Buck after running these two subcommands (but
  not for other subcommands), to save memory.

Since this is a hotfix,  I decided to hardcode in OCaml that `analyze` and
`infer` command always use `kill_buck_after_build=true`, regardless of what
Pyre client parses / produces. Otherwise it requires a release of Pyre client
(in addition to a release of an OCaml binary) to make sure the flag is set to
true for these two subcommands. This is not quick enough to mitigate the SEV.

Reviewed By: arthaud

Differential Revision: D68848222

fbshipit-source-id: b9377e1c1138260fe0f0979a3a730fbefe572b4d