Split the `Simple` case of `Attribute`.
Summary:
To handle properties, we need new forms of the `Attribute` enum, which
makes good names much more important than they have been to date.

To that end, it actually makes sense to split the existing case
to better express what is going on here instead of wrapping a `Result`.

Reviewed By: ndmitchell

Differential Revision: D68897128

fbshipit-source-id: 6df365c68a5d4796a47aa62bef780b6c8ea0050c