Move set access resolution to `AnswersSolver`
Summary:
I realized as I was about to implement property access that although
we don't really need to move the `set` access into `AnswersSolver`
for *functionality* at this stage (we will once we support setters!),
we do still need to move it to get optimal error messages because
otherwise we wouldn't have access to the class name which is used
in error messages.

So, moving it first in an easy refactor.

Reviewed By: ndmitchell

Differential Revision: D68897123

fbshipit-source-id: bbae35756945b10738705a450622d086913f2e63