Combine more behavior into attribute setting
Summary:
At the stage of handling property setters, it is no longer possible to
continue pretending that a "set type" of closed form always exists - we have
to deal with that simple case as just one branch of a bigger match in which
it may be necessary to analyze a synthetic function call.

The preexisting helper `resolve_set_access` becomes impossible to use without
a complicated return type; the easiest way forward is to just inline both
it and the downstream check (of either using a bidirectional `self.expr` call
or a type-level-only `is_subset_eq` call) into one helper that takes an
`Either` type to represent the contexutal vs non-contextual cases.

The resulting helper is a bit big and we might want to refactor it later,
but it does the job and is easier to work with as we are getting the
property setter support in place.

Reviewed By: ndmitchell

Differential Revision: D68910938

fbshipit-source-id: 1912eac66dad07d9082921d620deae2ac66dd550