Incomplete `if` statements might be the last in a function
Summary:
When looking for last statements, we assumed the if was complete. Now check that, and if it isn't, add the `if` itself as something that must obey whatever type.

You could argue this returns `None`. But then you probably want a location? Not super clear. But this is at least going to not miss errors.

Reviewed By: stroxler

Differential Revision: D68951798

fbshipit-source-id: 01ac6e9728c05c43ebaa75a6e2ba62de8a094b0d