Adjust missing return to be on the function
Summary:
Given:

```
def f() -> int:
    pass
```

We want to give an error that you can fall off the end. Currently we put it on the pass line. The conformance test puts it on the signature line. Let's match what the conformance test does.

At the same time, move to recording the expressions that must be Never, so we can actually assert their type in a subsequent phase.

Reviewed By: migeed-z

Differential Revision: D68969811

fbshipit-source-id: a75ec5ca281137dc470538a857e2534ddbf7f923