Take `ModuleInfo` out of `Error`
Summary:
error_driven_development

I was working on `ModuleInfo` related stuffs, and curiosity got me on why type error struct `Error` wants a handle to `ModuleInfo` -- If we look at [how Pyre1 models type errors](https://www.internalfb.com/code/fbsource/fbcode/tools/pyre/pyre2/pyre2/bin/error/legacy.rs), it's just a dumb plain data struct that retains the path and the source ranges.

I *think* one reason why `Error` wants a `ModuleInfo` is that it only stores range info in its `TextRange` form. We know that `TextRange` is not printable by itself -- it needs the "module contents" part of the `ModuleInfo` to be convertible to a printable source range. The way it seems to work today is that such conversion was performed very lazily -- it only happens when a source range is explicitly queried from the `Error`.

This diff builds on the observation that such laziness is neither necessary nor desirable. I think it's cleaner to perform the `TextRange` to `SourceRange` conversion immediately when the `Error` is created, for the following reasons:
- The conversion is guaranteed to be only calculated once upon `Error` creation this way.
- If the conversion fails for some reason, we get a stack trace that points exactly to the corresponding `Error` construction, which is far more helpful for debugging than a query-time stack trace. In fact, to get the construction-time stack trace under lazy mode one has to resort to D64396195 where an eager conversion gets triggered intentionally -- but if we are converting on construction anyway then what's the point of the laziness?
- To make `Error`s sortable & dedupable, the error collector code contains a blob of custom ord/eq wrapping logic just to skip over the `ModuleInfo` part of the `Error` type. If we do the eager range conversion, then `ModuleInfo` no longer needs to be part of `Error`, which makes `Error` dumb plain data again which can be easily made comparable without all those custom logic.

Reviewed By: ndmitchell

Differential Revision: D68877899

fbshipit-source-id: fa0f9e01a7d2f8d48c84c12a77ac5966f4e32383