Refactor the code to expose return annotation
Summary:
I would like to modify the following codeblock which currently handles `NoReturn` and isn't currently accurate:

      let never = function_last_expressions(&body, self.config);
        if never != Some(Vec::new()) ..

such that it adds a binding that records the return annotation, along with the return expression so that we can check if an expression is never then verify it against the type annotation. We will also delete parts of the code (the part which erases expression information and appends "None" to the return types)

Moving this block below where `return_ann` was originally defined resulted in a large number of failures, probably due to some mutations that are happening. So we refactor the code to expose `return_ann`.

In the next diff, we will create new bindings and record the annotation along with the return expressions.

Reviewed By: ndmitchell

Differential Revision: D68990378

fbshipit-source-id: c3ca09644cc6b2cc84580a78333c42f32401beff