Write models to shared memory during model generation
Summary:
# Context
We are having memory issue during model generations, and we suspect that is due to storing all generated models in the ocaml heap and sending them from workers to the main process using unix pipes.

# Solution
Let's write models directly to shared memory during model generation.

This requires passing a data structure (`SharedModels.t`) which remembers which targets have models, so we can free models from the shared memory later.

Unfortunately, we also have to update all the steps between model generation and the fixpoint. For instance, the inference of special models for dataclasses and named tuples.

# Functional change
Most of the changes in this diff should not be visible to the user. There is one exception: since we write models directly to shared memory, we do not have the specific model for a given callable and query. More concretely, this means the `--dump-model-query-results` will now dump the final models, i.e for a given callable, we will provide the join of all model queries.

Reviewed By: tianhan0

Differential Revision: D68834917

fbshipit-source-id: 9e1e959bab73757ba9f0482a4aeaf8df042aa824