Set default global strictness to unsafe if not explicitly set
Summary:
Just realized I made a bad decision when I was building the Configurationless migration, which was to select `# pyre-strict` as the default strictness in a global .pyre_configuration if no explicit strictness was specified.

Looking back on it, it's not the behavior of the Pyre type checker, and it can cause some confusing behavior, so I'm not sure why I did it.

This diff fixes that and mirrors the actual behavior of Pyre when choosing the global default.

Reviewed By: LoxFalcon

Differential Revision: D69130456

fbshipit-source-id: 9eadea131e9d18786de32b137b8ddb5e88a39bd3